+++ src/olc.h
@@ -109,7 +109,7 @@ struct	editor_cmd_type
  */
 AREA_DATA *get_vnum_area	args ( ( int vnum ) );
 AREA_DATA *get_area_data	args ( ( int vnum ) );
-int flag_value			args ( ( const struct flag_type *flag_table,
+unsigned long long flag_value			args ( ( const struct flag_type *flag_table,
 				         char *argument) );
 char *flag_string		args ( ( const struct flag_type *flag_table,
 				         int bits ) );
